(BSR)[PRO] refactor: delete unused EACInformation component
