+++ pro/src/utils/__specs__/venueHasCollectiveInformation.spec.ts
