+++ pro/src/utils/venueHasCollectiveInformation.ts
